webServer into GHCSpecter.Driver. package dep adjusted
