+++ daemon/app/Main.hs
@@ -1,61 +1,25 @@
 module Main (main) where
 
-import Concur.Core (Widget, liftSTM, unsafeBlockingIO)
-import Control.Applicative ((<|>))
 import Control.Concurrent (forkOS)
 import Control.Concurrent.STM
-  ( TChan,
-    atomically,
+  ( atomically,
     newTChanIO,
     newTQueueIO,
     newTVar,
-    newTVarIO,
-    readTChan,
-    writeTChan,
   )
-import Control.Lens ((.~), (^.))
-import Control.Monad.Extra (loopM)
 import Data.Aeson (eitherDecode')
 import Data.ByteString.Lazy qualified as BL
 import Data.Maybe (fromMaybe)
-import Data.Text (Text)
-import Data.Time.Clock (getCurrentTime)
 import GHCSpecter.Config
   ( Config (..),
     defaultGhcSpecterConfigFile,
     loadConfig,
   )
-import GHCSpecter.Data.Assets qualified as Assets
+import GHCSpecter.Driver (webServer)
 import GHCSpecter.Driver.Comm qualified as Comm
-import GHCSpecter.Driver.Session qualified as Session
-import GHCSpecter.Driver.Session.Types
-  ( ClientSession (..),
-    ServerSession (..),
-    UIChannel (..),
-  )
+import GHCSpecter.Driver.Session.Types (ServerSession (..))
 import GHCSpecter.Driver.Worker qualified as Worker
-import GHCSpecter.Render (render)
-import GHCSpecter.Server.Types
-  ( ServerState (..),
-    emptyServerState,
-  )
-import GHCSpecter.UI.ConcurReplica.Run (runDefaultWithStyle)
-import GHCSpecter.UI.ConcurReplica.Types
-  ( IHTML,
-    blockDOMUpdate,
-    unblockDOMUpdate,
-  )
-import GHCSpecter.UI.Types
-  ( HasUIState (..),
-    UIState,
-    UIView (..),
-    emptyMainView,
-    emptyUIState,
-  )
-import GHCSpecter.UI.Types.Event
-  ( BackgroundEvent (RefreshUI),
-    Event (BkgEv),
-  )
+import GHCSpecter.Server.Types (emptyServerState)
 import Options.Applicative qualified as OA
 
 data CLIMode
@@ -87,63 +51,6 @@ optsParser =
   OA.info
     (OA.subparser (onlineMode <> viewMode) OA.<**> OA.helper)
     OA.fullDesc
-
--- NOTE:
--- server state: shared across the session
--- ui state: per web view.
--- control: per web view
-
-styleText :: Text
-styleText = "ul > li { margin-left: 10px; }"
-
-webServer :: Config -> ServerSession -> IO ()
-webServer cfg servSess = do
-  let port = configWebPort cfg
-  assets <- Assets.loadAssets
-  runDefaultWithStyle port "ghc-specter" styleText $
-    \_ -> do
-      uiRef <-
-        unsafeBlockingIO $ do
-          initTime <- getCurrentTime
-          let ui0 = emptyUIState assets initTime
-              ui0' = (uiView .~ MainMode emptyMainView) ui0
-          newTVarIO ui0'
-      chanEv <- unsafeBlockingIO newTChanIO
-      chanState <- unsafeBlockingIO newTChanIO
-      chanBkg <- unsafeBlockingIO newTChanIO
-      let newCS = ClientSession uiRef chanEv chanState chanBkg
-          newUIChan = UIChannel chanEv chanState chanBkg
-      unsafeBlockingIO $ Session.main servSess newCS
-      loopM (step newUIChan) (BkgEv RefreshUI)
-  where
-    -- A single step of the outer loop (See Note [Control Loops]).
-    step ::
-      -- UI comm channel
-      UIChannel ->
-      -- last event
-      Event ->
-      Widget IHTML (Either Event ())
-    step (UIChannel chanEv chanState chanBkg) ev = do
-      (ui, ss) <-
-        unsafeBlockingIO $ do
-          atomically $ writeTChan chanEv ev
-          (ui, ss) <- atomically $ readTChan chanState
-          pure (ui, ss)
-      stepRender (ui, ss) <|> (Left . BkgEv <$> waitForBkgEv chanBkg)
-
-    stepRender :: (UIState, ServerState) -> Widget IHTML (Either Event ())
-    stepRender (ui, ss) = do
-      let renderUI =
-            if ui ^. uiShouldUpdate
-              then unblockDOMUpdate (render (ui, ss))
-              else blockDOMUpdate (render (ui, ss))
-      Left <$> renderUI
-
-    waitForBkgEv ::
-      -- channel for receiving bkg event
-      TChan BackgroundEvent ->
-      Widget IHTML BackgroundEvent
-    waitForBkgEv chanBkg = liftSTM $ readTChan chanBkg
 
 withConfig :: Maybe FilePath -> (Config -> IO ()) -> IO ()
 withConfig mconfigFile action = do
